Minor fixes and start work on additional animations
-Battlefield type is now randomize properly
-Adding equal operators to UUID
-ObstacleGraphicsInfo can store two more names for animations

+++ lib/battle/RandGen.cpp
@@ -10,7 +10,7 @@
 #include "StdInc.h"
 #include "RandGen.h"
 
-void RandGen::srand(int s)
+void RandGen::srand(int64_t s)
 {
 	seed = s;
 }
@@ -20,13 +20,13 @@ void RandGen::srand(int3 pos)
 	srand(110291 * pos.x + 167801 * pos.y + 81569);
 }
 
-int RandGen::rand()
+int64_t RandGen::rand()
 {
 	seed = 214013 * seed + 2531011;
 	return (seed >> 16) & 0x7FFF;
 }
 
-int RandGen::rand(int min, int max)
+int64_t RandGen::rand(int min, int max)
 {
 	if(min == max)
 		return min;
